fix: Attempt to truncate the file before writing
The "w" file mode does not truncate the file resulting in file
corruption if the written file is shorter than the existing file.
First, attempt to open with "wt" as a truncating write.  If that open
fails (as some providers may not support), open with "w".  Second, use
the file channel on the output stream to truncate the file descriptor.