Enhance DynamicPermissionMgr to handle non-required permissions
Permissions such as whether the app can post notifications are not
required for the operation of the app.  They should be requested, but
if denied, they shouldn't cause the app to complain and be
non-functional.

Each permission's state is tracked, and the interface methods operate on
them as a group.

The write external storage permission is no longer granted under SDK 33,
so it's state is always 'granted' then.  Likewise, the permission to
post notifications is new in SDK 33 and is always 'granted' before then.