Bind the YubiKey to the activity
Create YubikeyViewModel to hold the discovered USB or NFC YubiKey.
Refactor YubikeyMgr to use the YubikeyViewModel to access the YubiKey
when opening a file.

Add a few second delay when stopping access via NFC to allow the user to
move the key away from the phone so it doesn't activate multiple times.