Add debugging for Yubikey communication
Add an explicit debug tag, Yubikey.TRACE, which if enabled logs the
Yubikey messages at TRACE level.  At TRACE level, the low-level protocol
messages are logged including passwords.

Add an explicit logger for the com.yubico.yubikit package so it can be
modified when the debug tags change.

Add a label to the file open screen if YubiKey debugging is enabled